Don't change redirect URI if server is disabled
When using spotipy2 in combination with an external http server on a
(public) https page, authentication will fail because `get_redirect`
rewrites the URI to http.

This commit changes the behavior of `get_redirect` in such a way that it
doesn't touch the redirect URI if the built-in server is disabled.