Add check for unsuccesful token retrieval
When trying to retrieve the token, Spotify might reply with an error
like {'error': 'invalid_grant', 'error_description': 'Invalid redirect
URI'} which results in an inexpressive KeyError in Token.from_dict.

This commit adds a check for the HTTP Status Code that raises a
SpotifyException including the response body in case of error.